Rewrite using asyncio

+++ Makefile
@@ -19,14 +19,15 @@ install-systemd:
 	test $(shell id -u) == 0 || (echo "You must be root to install service"; exit 1)
 	useradd --system --home-dir=/ --shell=$(shell which nologin) $(USER) || :
 	cp ${DIR}$(CODE_SRC) $(CODE_DST)
-	cp $(DIR)$(CONF_SRC) $(CONF_DST)
+	test -f $(CONF_DST) || cp $(DIR)$(CONF_SRC) $(CONF_DST)
 	chmod 755 $(CODE_DST)
-	sed 's|{{USER}}|$(USER)|; s|{{EXEC}}|$(CODE_DST) $(CONF_DST)|' $(DIR)$(SYSTEMD_SRC) > $(SYSTEMD_DST)
+	sed 's|{{USER}}|$(USER)|; s|{{EXEC}}|$(CODE_DST) --syslog $(CONF_DST)|' $(DIR)$(SYSTEMD_SRC) > $(SYSTEMD_DST)
 	systemctl daemon-reload
 
 erase-systemd:
 	test $(shell id -u) == 0 || (echo "You must be root to erase service"; exit 1)
 	systemctl stop $(SERVICE_NAME) || :
+	systemctl disable $(SERVICE_NAME) || :
 	rm -f $(SYSTEMD_DST)
 	rm -f $(CODE_DST)
 	rm -f $(CONF_DST)
